requirements.txt
# Python Dependencies for Silver Tier AI Employee

# Core dependencies
python-dotenv==1.0.0

# Gmail Watcher
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0

# WhatsApp Watcher (Playwright)
playwright==1.40.0

# Filesystem Watcher
watchdog==3.0.0

# Logging and utilities
python-dateutil==2.8.2
